+++ templates/button.dart
@@ -1,8 +1,8 @@
-// TODO: import 'styling.dart';
-// TODO: import 'spinner.dart';
-
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
+
+import 'spinner.dart';
+import 'styling.dart';
 
 enum ButtonVariant { primary, secondary, ghost, destructive, outline, link }