+++ templates/nav_bar.dart
@@ -1,7 +1,7 @@
-// TODO: import 'styling.dart';
-
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
+
+import 'styling.dart';
 
 class NavBarItem {
   final Widget icon;